final compatabilites with new api

+++ src/main/java/org/yesworkflow/recon/UriVariableValue.java
@@ -2,11 +2,11 @@ package org.yesworkflow.recon;
 
 public class UriVariableValue
 {
-    public Integer uriVariable;
-    public Long resource;
+    public Long uriVariable;
+    public Integer resource;
     public String value;
 
-    public UriVariableValue(Integer uriVariable, Long resource, String value)
+    public UriVariableValue(Long uriVariable, Integer resource, String value)
     {
         this.uriVariable = uriVariable;
         this.resource = resource;
